% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scMTD.r
\name{scMTD}
\alias{scMTD}
\title{A statistical multidimensional imputation method for single-cell RNA-seq data.}
\usage{
scMTD(data, do.nor = TRUE, do.log = TRUE, cores)
}
\arguments{
\item{data, }{Gene expression matrix (gene by cell).}

\item{do.nor, }{Normalized step.Default is TRUE.}

\item{do.log, }{logarithm transformation step.Default is TRUE.}

\item{cores, }{A integer specifying the number of cores used for parallel computation.}
}
\value{
An imputation matrix
}
\description{
A statistical multidimensional imputation method for single-cell RNA-seq data.
}
\examples{
library("scMTD")
data(data)
imputed_data<-scMTD(data,do.nor=TRUE,do.log=TRUE,cores=2)
}
